backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pandas
python-dotenv
pydantic
